Added a new item page
The new item page allows users to add a previously unknown item to the
inventory. While the same functionality can be achieved with a restock
page, this provides a clear interface for the action.

Additionaly fixed a bug where client-side scripts wouldn't load if
Handlebars wasn't defined. Also fixed the wrong error message and status
code being returned when no item was specified when adding an item to
the inventory.