Added items to order page
To inform the user of which items are below minimum stock and need to be
ordered, created items to order page. It shows the item's name,
wholesale price and how many units to order to reach maximum stock
levels. It also shows the total order price for items individually and
an aggregated total at the footer of the table.

- Added a new Handlebars page.
- Created a new method in the module to extract items below minumum
stock and map their order count and price.
- Added unit tests testing the new functionality.