Resolved issue#20: Incorrectly displayed price
Resolves issue #20. https://github.coventry.ac.uk/340CT-2021OCTJAN/jansonsa-sem1/issues/20
Item was inserted into the HTML manually, which required the price to be
changed manually which wasn't done.

Fixed by using client-side Handlebar templates to display the item. This
approach can use the already created "asPounds" Handlebars helper which
displays pence as pounds. Doing this also removed multiple lines of code
that are no longer used.