Added item details page
The website was lacking the ability to view and modify details of
individual items, which made interacting with the website  more
complicated.

Created an item details page, which displays all the item's details in a
form, allowing any and all of the fields to be updated. This page can be
accessed by clicking on the item's name from either "All items" page or
"Items to order" page.