Extracted header as partial and new hamburger menu
The initial design was very unfriedly for mobile devices as all items
were sized according to the screen's width, making the page on mobile
unusable.

Updated the design to be mobile-first, by using a hamburger menu that
expands. Additionally updated the header size to use REM instead of VW.
Improved the header design. Created a simple icon, so that the page
doesn't use the Google logo as an icon.