Extracted the head of pages into a partial
There was a lot of common information in the head of each page,
including the charset, author and stylesheets. It also had
elements that only changed their values: title and description.
Adding a new style sheet or a common script would require all pages to
be updated. Having the head in a single place would make the code easier
to maintain.

Extracted this common part into its own Handlebars partial taking
parameters for title and description. Left the wrapping <head> element
in each of the pages so that individual scripts can still be added where
necessary, without them being used in every page.