Added Footer to all pages
None of the pages had a footer, which is usually a very convenient
element containing lots of important information.

Created a footer as a Handlebars partial and used it in all pages to
display a footer. It is fairly simple, just displaying the copyright
symbol, year and author's name. To position the footer so that:
- When a page is scrollable, it is at the very bottom;
- When a page is not fully filled, it is at the bottom of the viewport.
I achieved this by using a flex display on the Body element with a
column direction and setting the min width of the body to be 100wh. Then
I positioned the footer using a margin-top: auto so that it is displayed
at the bottom of the body.