Created section style and added section headers
By default all sections were overflowing and had a very basic style.
When sections were empty, they would still appear as blobs on the page
and confuse users.

Styled sections to fit the overall design and appear as white cards.
Added a header to all sections that gives context about the section,
which is particularly helpful in pages that have multiple sections.
Styled these headers to be a lighter color and have a separator line.
Added sections to elements that still didn't have them as Login and
Register.