Align sections using Flex container
By default all the sections appeared below each other which led the user
to do lots of scrolling to take in all of the information.

Since the information doesn't take up much space horizontally, styled
the main container to have display flex and the sections inside to have
flex grow properties, allowing them to be next to each other and
wrapping when the screen becomes small enough. Also styled the main
container to have a max width, so that on widescreen monitors the
sections are in the center instead of being overly stretched out.