Created route to upload images
Currently item images were passed as links. This meant that the image
had to be on the internet so that it could be uploaded.

Created an images endpoint which accepts image files and stores them on
the machine in /public/images directory. Ideally the images should be
stored on a file store like S3, but for prototyping purposes this should
be a good intermediate solution. Relying on koa-body library, which puts
the images into ctx.request.files and they can be copied for persisted
storage. The upload route returns the image's location on the server and
it can be accessed by this link on the internet. This link can then be
used to set it as the photo field for items. Additionally, created an
images directory in the public folder, so that the application doesn't
need to create it and added a .gitkeep so that it is included in version
control. All uploaded files will be ignored from version control through
.gitignore.