Uploading images when creating new items
Until now there was no way to upload an image when creating a new item.

A new item can be created from either the restock page or the new item
page. Created a script to upload the file, display it and put the
uploaded images URL into the "photo" form field. Reused this script in
both pages to add the functionality. Changed the item details page to
allow the picture URL to be manually changed. Added a style that hides
images that have an empty src attribute.