Created a script to populate the inital table
SQLite is a local file database and therefore doesn't persist data if
the local file is deleted. Heroku does not guarantee local file
persistance and when the website is run on Heroku, we can't ensure that
is has any data at any point.

I created a local script to create the tables and populate them if they
dont exist. I then run this script after every npm install, as Heroku
runs npm install before starting the server.

I added more items to the items table and used my university email
address for accounts and as the shop owner.