Added stock level column to all items page
Items would only emit an order notification when their stock fell below
minimum stock level. If an item was near the minimum stock level there
was no easy way to know this.

Added 2 extra columns to the all items page:
- First column shows the minimum and maximum stock level.
- Second column shows the current stock level as percentage relative to
minimum stock. For example, if an item's minimum stock was 10, the
maximum stock was 20 and the current stock was 13, then the stock level
would be shown as 30%. At 100% it would be the same as maximum stock and
at 0% it would be the same as minimum stock. This number can go over 100
to indicate that there is too much of a particular item and go negative,
indicating that an item needs ordering.