Remember default Restock item placeholder
Restock page had a placeholder section instructing the user to scan an
item, however when restock was reset, the HTML in this section was
reset.

Remember the inner HTML of the item element by extracting it when DOM is
loaded and using it when the page is reset. Also displaying the image
in every item's detail page.